fix(edit): fix message edits loses the album of images
Fixes #16741

The problem was that we replaced the ContentType from Image to Text, so on restart, it doesn't understand it had images anymore.

Fixed in status-go by reusing the ContentType of the original message.

This makes it so that we don't need to pass the ContentType from Nim anymore, so I removed that param from the code.