fix(block): remove chat and messages when blocking a contact (#16889)
Fixes #16640

This makes it so that when you block a contact, it now also removes the chat and the messages as expected by the requirements and as Mobile does.

To do so, I use the same API as mobile instead of the forked desktop one. I removed the desktop one as it is no longer needed (see status-go PR)

I also fixed an issue when unblocking where it would send a double toast messages with one saying you "removed the contact", but it was already removed.