fix(ac): fix review CR popup not showing correct data from the AC
Fixes #16601

The original issue of seeing `&nbsp;`  was already fixed, but the popup, when opening from the activity center didn't have any spaces.
However, when displayed from the profile, everything showed fine.
So the fix was simply to remove the old way of opening from the AC and use the same call that is used in the profile popup.