feat(onboarding): add AC notif for importing old accounts (#17078)
* feat(onboarding): add AC notif for importing old accounts

Fixes #17028

When an old user imports an account, we now fetch the backups in the background and show an AC notification.

When the fetch is successful, the AC notif switches to a success message.

If after a timeout we detect that we didn't fetch anything or just part, we show an error and the possibility to try again.

* fix: typo

* chore: bump status-go

* fix: bump status-keycard-go

---------

Co-authored-by: Igor Sirotin <sirotin@status.im>