StatusQ: StatusQrCodeCapture component excluded from StatusQrCodeScanner
Excludes StatusQrCodeCapture which is responsible exclusively for qr
codes detection from the video input, with no other UI elements. It's
the only component related to qr codes depending directly on QZXing and
QtMultimedia. That depenency is removed StatusQrCodeScanner. It gives
ability of providing alternative version on the StatusQrCodeCapture for
qt 6 via file selector.