fix(wallet): stop previous route updates if there is an input validation error
If the initial route was successful and updates started for it, after the user changes some input route
params the app will try to generate a new route. While updating if a local error occurs (like not enough
ETH or assets) the app will not try to generate a new route, while the updates for the previous route
won't be stopped even that's not a valid route anymore. This commit stops any updates of the previous
route.