perf(login): improve login time by skipping backend calls for unknown contact
Fixes #17303

It's unclear the reason why the login time degraded so much from 2.32 to 2.33, since the actual code didn't change for what I fixed here.
The problem I uncovered is that we have a lot of direct backend calls (sync) in the contact service for unknown contacts. The reason is that for unknown contacts, we just have the pubkey, so we need to get the colorID, the colorHash and the alias.
For small calls like updating a contact, checking a single contact and such, it's fine. However, when we open the app, we generate the member lists we are a part of or that are in the curated list.

The latter is easy to fix. We actually do not need the member details, so I just populate the member list with empty members so that the number of members still works, but it stays fast and doesn't use memory.

For joined communities, it's harder to circumvent, because we actually need the member details for the admin member list and the channel lists.
In that case, I decided to just not call the backend and let the frontend call it, since our QML code is now smart and only generates the Member list items for what is on screen, so we will only call the backend for the couple of visible members, instead of the thousands in the Status community.

We can still improve it, but we can do that in further commits, that way we have a immediate remedy for the super long logins.