fix: simplify nim finalizers
Remove any extra work done in the finalizers, keeping just the nimqml `delete` that are needed to clean the c++ classes.

Apparently the finalizers are buggy and we need to be careful not to allocate any memory in it.

Motivation: nim-lang/Nim#4851
(cherry picked from commit 5ffac3cdf40d51bfb7e3bbecf58e5bb3dbb09f40)