fix(login): move to main only when the chat is loaded
Fixes #17629

Only moves to the main screen when the chats are loaded to avoid showing the "loading sections" message.
Also removes the loadingSection as it will never have a chance to be shown again.
Removes the 3 second debounce when switching to main as the loading should be all finished by that point anyway