fix: Provide a general safe decoding for Json
This is a follow-up of https://github.com/status-im/status-desktop/pull/17594 that's closing the door for the json decode issues on all usages and all platforms.

There's been another attempt here https://github.com/status-im/status-desktop/pull/17606, but it's not efficient on Windows and other platforms. It only works on MacOs asa far as I can tell.