fix(news): fix toast notifs sometimes showing when they should not
Fixes #17895

The issue was that I was too permissive when I wrote the code to show the toasts and it triggered any time a notification was added or loaded, so on app start, it could trigger too.
However, I got lucky because most of the time the load happened before the NotificationManager was ready, so this race condition hid the issue.

I fixed it by using different event names for when an AC notif is received vs loaded