fix(StatusBaseInput): clicking in bottom text of the input not working
Fixes #14467

The issue was that the TextEdit was bound to the size of the flickable, so when it got scrolled, mouse clicked we only on the ScrollView and no longer on the TextEdit, ie the mouse click was only good to scroll (drag) but not to select text.

Removing the height property and letting the TextEdit follow its content fixes it.