fix(tokens): fix token holder list not showing the right member
Fixes #17146

The issue was that we didn't use the LeftJoinModel correctly. The TokenHolderModel should have been on the left.

Then, we often do not have access to the shared address of the members as a TokenMaster, so it would show an empty member.
I fixed that by doing a clean up so that we use the normal token list component, which has a mode where it shows the address only if we don't have access to the member's info.

This differs from the design a little but makes the code way cleaner, as we don't need two different components and also I removed the use of `getContactbyIDJson`