fix(tokens): fix own owner token not showing details
This is a workaround fix to the problem that we do not store our own revealed addresses as an Owner.

This means that we did not show the owner details in the token holder list.

The fix here is to simply identify if we got the Owner token holders and if we are the owner, we set the Pubkey to ourselves. This is sufficient to get the full details from the member list during the JoinModel