fix(shell): fix chats needed to be loaded and images not showing
Fixes #18030

I moved the initialization of the chats to the Nim side. The personal section's chats get loaded on app start now. It will impact the login a little, but not much, since it only loads the model of chats, not messages.

For the image issue, the problem was that the ShellAdaptor didn't have the `chatType` as an `expectedRole`. Same for `onlineStatus`. I guess there is a bug in the SFPM where without the expectedRole, it's undefined at first, but then when you reload the page it works, the bug being that it somehow works on reload.