fix(Storybook): unbreak (hot) reloading the pages
- it's been a bit flimsy under Qt 6.9, and completely stopped working in
6.9.1
- simplify the QML logic for reloading the page: guard the reloading
process with a simple boolean flag, instead of working directly on the
view loader (suggested by Pepe in
https://bugreports.qt.io/browse/QTBUG-135448?focusedCommentId=882898&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-882898)

- improve the C++ cache cleaner with first manually running the GC, send
a delayed event to all objects to do a cleanup/delete, and then finally
clear the component cache

Quoting:
> clearComponentCache() indeed does another GC run. However, both GC
runs are required here. The first one drops the component's JavaScript
wrapper, calling deleteLater() on the component. Then the event
processing deletes the component and instance QObjects. The second GC
run drops the compilation unit the component and the object were holding
on to.

Fixes https://github.com/status-im/status-desktop/issues/18069