fix(profile_image): show the contact icon when no display name is set
Fixes #17487

When no displayName, nickname, ens name or user image are set, we show the contact icon instead, in accordance to the design.

Also fixes an issue with the SuggestionBox where the colorHash was not set correctly.
Then fixes an issue where right-click on contact items didn't work

Finally, this refactors some parts of the code to use UserImage instead of SmartIdenticon, as it simplifies the code and its maintenance