fix(images): fix sending pasted images
Fixes #18181

There were two problems
First, we added a path formatting function in the QML that just deleted the image data. Making it so that we sent nothing.
I fixed that and also improved the error management so that when we send nothing, we show an error to the user instead of leaving them stuck.

Second, we removed the resize function that also put the image data in a temp file. So now the status-go code couldn't handle the data images.
I fixed that by re-adding part of the resize function that puts the data inside a team file.