chore(wallet): necessary changes due to router simplifications on the statusgo side
- Since the app operates in a single chain environment, there is no need for maintaining selected chains on from/to sides
as arrays of disabled chains. Replaced `disabledFromChainIDs and `disabledToChainIDs` with `fromChainID` and
`toChainID` parameters across multiple files. And code around aligned with the change.
- Removed unused functions related to disabled chain IDs.
- Updated UI components to reflect the new parameter structure for route suggestions.

Old SendModal which is now used only for the Bridge transactions is updated in a way that:
- "Simple" view is removed (only advanced remained), as well as the tab bar, cause there is a single option now.
- For "Advanced" view the logic for enabling/disabling more chains is updated so that users enable a single chain
by clicking on it, while all others get disabled (this applies for both, from and to chains). This way, a single chain
selection is enforced on the UI side until a new UI for the Bridge flow gets implemented.