perf(chat_section): lazy load chat modules and improve performance
Fixes #18361 and #14431

Lazy loads the chat section's children modules so that only when a chat is activated are the modules loaded (content, users, messages and chat input).
This improves performance and memory a little bit.

Also fixes other performance issues in the chat_section for communities, like useless calls to the service, useless calls to update the model (for categories) and also useless calls to update the badge.
All of those slowed down the load. Most of those were fixed by simply accessing the `community` object already available.