perf(main): lazy load section members when going to the admin screen
Fixes #18370

We were generating all the members items for communities on app start even for normal users that would never have access to it. Indeed, that model is only shown when accessing the admin panel.
Hence, we now wait until the user goes to the community Admin screen to load the members.

We also lazy load shared addresses and community tokens, because, again, they are only shown in the admin panels.