fix(community): fix user not shown self as banned in community
Fixes #18412

The problem was introduced when I implemented the lazy loading for the members in the section model. Turns out even if that member list is never shown to members, we do use it for some checks, like the banned one.

I fixed by passing the banned value to the sectionItem directly. Relying on the members model was a crutch.