switch regex library
`std/re` is more or less deprecated as it relies on `pcre` which itself
is EOL.

Here, we replace all (3!) usages with
[`nim-regex`](https://github.com/nitely/nim-regex), which has the added
benefit that parsing the regex is done at compile time and therefore
would result in a compile-time error if the regex is bad.

This changes allows dropping pcre from the build, at least as far as the
nim parts are concerned.