chore: backport refc lifetime fix to nim-serialization
https://github.com/status-im/status-desktop/pull/17766 introduced a
workaround for a refc issue in Nim - this PR moves the "fix" to
`nim-serialization` proper to extend it to all applications.

With the nim-serialization bump, the changes in #17766 can be reverted.

Included is also a bump of `json_serialization` which includes code for
efficient streaming as well as bug fixes.