fix(contacts): fix unknown contacts not having an alias
Fixes #18548

The problem was that we introduced a mechanism to speed up login by skipping backend calls on contacts on app start. This was fine before because we also relied on the service to get the contact details.
However, we now use the contacts model instead, which is faster. So that was empty. Plus, we were storing an empty alias in the contacts table, so the contact never populated.

Now, the fix is simple, we stop using `skipBackendCalls`, because we already stopped loading any member lists on app start, so we no longer need that.
Also, we update the contacts model when we save a new contact, making sure there is an alias for them.