fix(perf): fix freeze on opening big communities
Fixes #18636

The logs were a red herring, but also it showed that we had some useless code, which I removed.
The real fix was to remove the event to CONTACT_UPDATED. That call was not really necessary and slowed the app a whole lot, because it got called a lot and is listened to by multiple modules.