fix(contacts): populate contacts model for unknown contacts
Found while fixing #18636

Since we use ModelEntry to get the contact details of users now, for the ProfilePopup for example, some contacts just do not have any data. That's because we never got their contact data from the store nodes, so we can only generate their alias from the pubkey.

Since the contact model is initially generated from the contacts cache, it means that if you want to look at the Profile of an "unknown" user, it will not exist in the model.

The fix is to populate the model with a generated contact, by calling the module function. I needed to add a function that asks the model first if it has the user, then we call the module to populate it and the data changed is called automatically, so it's all instant to the end user.