fix(syncing): don't show the syncing AC notif if the pairing worked
Fixes #18633

We weren't checking to see if the pairing failed, only that a paring happened.

I think the regression happened during the refactor. That code was probably in a code path that could only be reached by restoring using a mnemonic and now we only have one "ending" path (simpler).