ci: validate secrets after bundling
Introduces --verify-credentials flag which can be passed to status app and it will check for all the necessary credentials provided by status-jenkins-lib.
related changes are here : https://github.com/status-im/status-jenkins-lib/pull/120

Adds an Audit stage in jenkins which checks for these credentials, this is an integration test in CI to ensure that bundled app has all the necessary secrets.

Shorten existing stage names in Jenkins for better readability.