fix: Wallet crashing on deleted QObject access in QML
This commit fixes the crashes produced by temporary `CurrencyAmount` objects shared to QML via slots.
This update will store the `CurrencyAmount` in the QObject that exposes the slots and the QObject will live as long as the QObject parent is alive.

The exception here `walletSection.getCurrencyAmount`. Now the API will return a json string.