chore: Add splash screens
+ remove the safe area workarounds for qt 6.9.1

This commit adds the splash screen resources for Android and IOS.
On Android it will also create a StatusBarUtil.java. This is a helper to notify the OS when the app theme changes so that the top bar icons to change as well.
StatusQtActivity class inherits the QtActivity and implements the splash screen condition on Android. This is necessary to avoid screen flickering and empty screens while the app loads.