Poc: keyboard native feel on android (#19139)
* chore: resize the view whenever the keyboard shows up

Also use a fullscreen popup on mobile when a close button is available.

* fix: Exclude from fullScreenSheet the popups that won't behave nicely

Fixing the authentication popup overlapping the header. This one is a must

* fix: Manually resize the view when the keyboard shows up

Android: Setting  `android:windowSoftInputMode="adjustNothing"`. Introducing keyboardUtil helper to provide the keyboard size because qt won't provide it consistently
IOS: Implementing a frame tracker monitor qt attempts to move the window when showing the keyboard. The attempt will be cancelled and the window resize will be handled manually in qml

main.qml: Account for the keyboard movements and add to the safe areas

statusModal.qml: Fix hasCloseButton property
statusBarUtil.java: Fix the OS icons/buttons color