fix(backup): fix Android backups by using SAF
Fixes #19082

To fix the issue, we need to use the Storage Access Framework (SAF), meaning that Go cannot just write to disk. It needs to first write to the sandbox, then using native code, we can move that file from the sandbox to the allowed file.

This makes the backup work as expected.