fix(account): fix account rename being reset by binding
Fixes #19174

The problem was that the popup was being binded to the selected account, so when a signal was triggered in the background from an async task, it would retrigger the binding and reset the popup.

Passing the account properties directly is cleaner and fixes the issue.

I also renamed the Component because there was a type (missing `u`)