fix(settings): fix the naming of "Allow new contact requests" setting
The setting was named "Allow new contact requests", but the setting itself is named `MessagesFromContactsOnly`, so turning it off affected all messages.

We think that this is an old setting that dated to when we didn't have communities, but now that we do have them, it also stops receiving messages from non-contacts in the community as well.

Renaming the label makes it clear what the setting does.