fix: Make sure the nfc channel can be opened at start-up
If the app needs the nfc channel at startup qt will enter in a race-condition. The result is that qt cannot enable the reader since the java infrastructure-setup is not ready for qt. The nfc can be enabled afterwards only by placing the app in the background and then back to foreground.

To workaround this issue we'll make sure the foregroundDispatch gets enabled at startup.